Rename folder, re-fit/re-frame components
