+++ src/client/src/components/ProfilePage/ChallengeMiniCard.vue
@@ -2,28 +2,32 @@
   <div class="container-fluid p-0 text-light">
     <section class="row border-underline background-highlight  align-items-center pt-2"
       @click="openChallenge(challenge.id)" v-if="challenge">
+      <div class="col-11 p-0">
+        <section class="row">
+          <div class="col-12 col-lg-6">
+            <p class="m-0 fw-bold fs-5 text-truncate">
+              {{ challenge?.name }}
+            </p>
+          </div>
 
-      <div class="col-12 col-md-6 col-lg-5">
-        <p class="m-0 fw-bold fs-5 text-truncate">
-          {{ challenge?.name }}
-        </p>
+          <div class="col-8 col-lg-4">
+            <span class="fw-semibold text-uppercase">
+              {{ challenge.category }}
+            </span>
+          </div>
+
+          <div class="col-4 col-lg-2">
+            <span class="highlight-font fw-semibold fs-5">
+              <i class="mdi mdi-diamond-stone me-2 fs-5" :style="{ color: difficulty.color }"></i> {{ challenge.difficulty
+              }}
+            </span>
+          </div>
+        </section>
       </div>
 
-      <div class="col-8 col-lg-3">
-        <span class="fw-semibold text-uppercase">
-          {{ challenge.category }}
-        </span>
-      </div>
-
-      <div class="col-4 col-lg-2">
-        <span class="highlight-font fw-semibold fs-5">
-          <i class="mdi mdi-diamond-stone me-2 fs-5" :style="{ color: difficulty.color }"></i> {{ challenge.difficulty }}
-        </span>
-      </div>
-
-      <div class="col-6 col-lg-1 my-2 d-flex">
+      <div class="col-1 my-2 d-flex justify-content-md-center">
         <i @click.stop="leaveChallenge()" :title="`Leave challenge: '${challenge.name}'`"
-          class="btn mdi mdi-file-cancel-outline m-auto text-muted"></i>
+          class="mdi mdi-file-cancel-outline text-muted"></i>
       </div>
 
     </section>
