+++ src/client/src/components/ProfilePage/StatCard.vue
